requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
click>=8.1.0
rich>=13.0.0
streamlit>=1.35.0
